requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data processing
pandas==2.1.3
duckdb==0.9.2
pyarrow==14.0.1  # For Parquet support
openpyxl==3.1.2  # For Excel file parsing (Vyapar exports)

# Database
aiosqlite==0.19.0  # Async SQLite

# AI services
openai==1.3.5
anthropic==0.7.7  # Claude

# PDF processing
pypdf2==3.0.1
pdfplumber==0.10.3

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Background jobs
celery==5.3.4  # Optional, can use simple queue
redis==5.0.1  # If using Celery

# Validation and hashing
cryptography==41.0.7

# Logging
structlog==23.2.0

# Testing (dev)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # For testing FastAPI

# Code quality (dev)
black==23.11.0
flake8==6.1.0
mypy==1.7.0
